Moved the zoom/pan logic up into the base viewer
