+++ glue_vispy_viewers/volume/volume_viewer.py
@@ -67,73 +67,6 @@ class VispyVolumeViewer(BaseVispyViewer):
         # FIXME: needs to be done after first layer added
         # self._update_clip(force=True)
 
-    def camera_mouse_wheel(self, event=None):
-
-        scale = (1.1 ** - event.delta[1])
-
-        with delay_callback(self.state, 'x_min', 'x_max', 'y_min', 'y_max', 'z_min', 'z_max'):
-
-            xmid = 0.5 * (self.state.x_min + self.state.x_max)
-            dx = (self.state.x_max - xmid) * scale
-            self.state.x_min = xmid - dx
-            self.state.x_max = xmid + dx
-
-            ymid = 0.5 * (self.state.y_min + self.state.y_max)
-            dy = (self.state.y_max - ymid) * scale
-            self.state.y_min = ymid - dy
-            self.state.y_max = ymid + dy
-
-            zmid = 0.5 * (self.state.z_min + self.state.z_max)
-            dz = (self.state.z_max - zmid) * scale
-            self.state.z_min = zmid - dz
-            self.state.z_max = zmid + dz
-
-        self._update_clip()
-
-        event.handled = True
-
-    def camera_mouse_press(self, event=None):
-
-        self._initial_position = (self.state.x_min, self.state.x_max,
-                                  self.state.y_min, self.state.y_max,
-                                  self.state.z_min, self.state.z_max)
-
-        self._width = (self.state.x_max - self.state.x_min,
-                       self.state.y_max - self.state.y_min,
-                       self.state.z_max - self.state.z_min)
-
-    def camera_mouse_release(self, event=None):
-        self._initial_position = None
-        self._width = None
-
-    def camera_mouse_move(self, event=None):
-
-      if 1 in event.buttons and keys.SHIFT in event.mouse_event.modifiers:
-
-          camera = self._vispy_widget.view.camera
-
-          norm = np.mean(camera._viewbox.size)
-
-          p1 = event.mouse_event.press_event.pos
-          p2 = event.mouse_event.pos
-          d = p2 - p1
-
-          dist = (p1 - p2) / norm * camera._scale_factor
-          dist[1] *= -1
-          dx, dy, dz = camera._dist_to_trans(dist)
-
-          with delay_callback(self.state, 'x_min', 'x_max', 'y_min', 'y_max', 'z_min', 'z_max'):
-
-              self.state.x_min = self._initial_position[0] + self._width[0] * dx
-              self.state.x_max = self._initial_position[1] + self._width[0] * dx
-              self.state.y_min = self._initial_position[2] + self._width[1] * dy
-              self.state.y_max = self._initial_position[3] + self._width[1] * dy
-              self.state.z_min = self._initial_position[4] + self._width[2] * dz
-              self.state.z_max = self._initial_position[5] + self._width[2] * dz
-
-          event.handled = True
-
-
     def mouse_press(self, event=None):
         if self.state.downsample:
             if hasattr(self._vispy_widget, '_multivol') and not self._downsampled:
